TS Move common types to core package #9779
Moved Aggregations and Filters to core.
