+++ modules/lib/lib-node/src/main/resources/lib/xp/node.ts
@@ -13,7 +13,17 @@ declare global {
     }
 }
 
-import type {ByteSource, PrincipalKey, QueryDsl, SortDsl} from '@enonic-types/core';
+import type {
+    Aggregation,
+    AggregationsResult,
+    ByteSource,
+    Filter,
+    Highlight,
+    HighlightResult,
+    PrincipalKey,
+    QueryDsl,
+    SortDsl,
+} from '@enonic-types/core';
 
 export type {
     ByteSource,
@@ -34,6 +44,32 @@ export type {
     BooleanDslExpression,
     MatchAllDslExpression,
     PathMatchDslExpression,
+    DateBucket,
+    BucketsAggregationResult,
+    StatsAggregationResult,
+    SingleValueMetricAggregationResult,
+    AggregationsResult,
+    Aggregation,
+    TermsAggregation,
+    HistogramAggregation,
+    DateHistogramAggregation,
+    NumericRange,
+    NumericRangeAggregation,
+    DateRange,
+    DateRangeAggregation,
+    StatsAggregation,
+    GeoDistanceAggregation,
+    MinAggregation,
+    MaxAggregation,
+    ValueCountAggregation,
+    Highlight,
+    HighlightResult,
+    ExistsFilter,
+    NotExistsFilter,
+    HasValueFilter,
+    IdsFilter,
+    BooleanFilter,
+    Filter,
   } from '@enonic-types/core';
 
 type WithRequiredProperty<T, K extends keyof T> = T & { [P in K]-?: T[P] };
@@ -43,215 +79,6 @@ function checkRequired<T extends object>(obj: T, name: keyof T): void {
         throw `Parameter '${String(name)}' is required`;
     }
 }
-
-// START AGGREGATIONS, FILTERS, QUERIES, SUGGESTIONS
-
-export interface Bucket {
-    [subAggregationName: string]: AggregationsResult | string | number | undefined;
-
-    key: string;
-    docCount: number;
-}
-
-export interface NumericBucket
-    extends Bucket {
-    from?: number;
-    to?: number;
-}
-
-export interface DateBucket
-    extends Bucket {
-    from?: string;
-    to?: string;
-}
-
-export interface BucketsAggregationResult {
-    buckets: (DateBucket | NumericBucket)[];
-}
-
-export interface StatsAggregationResult {
-    count: number;
-    min: number;
-    max: number;
-    avg: number;
-    sum: number;
-}
-
-export interface SingleValueMetricAggregationResult {
-    value: number;
-}
-
-export type AggregationsResult = BucketsAggregationResult | StatsAggregationResult | SingleValueMetricAggregationResult;
-
-export type Aggregation =
-    | TermsAggregation
-    | HistogramAggregation
-    | DateHistogramAggregation
-    | NumericRangeAggregation
-    | DateRangeAggregation
-    | StatsAggregation
-    | GeoDistanceAggregation
-    | MinAggregation
-    | MaxAggregation
-    | ValueCountAggregation;
-
-export interface TermsAggregation {
-    terms: {
-        field: string;
-        order?: string;
-        size?: number;
-        minDocCount?: number;
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface HistogramAggregation {
-    histogram: {
-        field: string;
-        order?: string;
-        interval?: number;
-        extendedBoundMin?: number;
-        extendedBoundMax?: number;
-        minDocCount?: number;
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface DateHistogramAggregation {
-    dateHistogram: {
-        field: string;
-        interval?: string;
-        minDocCount?: number;
-        format: string;
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface NumericRange {
-    from?: number;
-    to?: number;
-    key?: string;
-}
-
-export interface NumericRangeAggregation {
-    range: {
-        field: string;
-        ranges?: NumericRange[];
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface NumericRangeAggregation {
-    range: {
-        field: string;
-        ranges?: NumericRange[];
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface DateRange {
-    from?: string;
-    to?: string;
-    key?: string;
-}
-
-export interface DateRangeAggregation {
-    dateRange: {
-        field: string;
-        format: string;
-        ranges: DateRange[];
-    };
-    aggregations?: Record<string, Aggregation>;
-}
-
-export interface StatsAggregation {
-    stats: {
-        field: string;
-    };
-}
-
-export interface GeoDistanceAggregation {
-    geoDistance: {
-        field: string;
-        unit: string;
-        origin?: {
-            lat: string;
-            lon: string;
-        };
-        ranges?: NumericRange[];
-    };
-}
-
-export interface MinAggregation {
-    min: {
-        field: string;
-    };
-}
-
-export interface MaxAggregation {
-    max: {
-        field: string;
-    };
-}
-
-export interface ValueCountAggregation {
-    count: {
-        field: string;
-    };
-}
-
-export interface Highlight {
-    encoder?: 'default' | 'html';
-    tagsSchema?: 'styled';
-    fragmenter?: 'simple' | 'span';
-    fragmentSize?: number;
-    noMatchSize?: number;
-    numberOfFragments?: number;
-    order?: 'score' | 'none';
-    preTag?: string;
-    postTag?: string;
-    requireFieldMatch?: boolean;
-    properties?: Record<string, Highlight>;
-}
-
-export interface HighlightResult {
-    [highlightedFieldName: string]: string[];
-}
-
-export interface ExistsFilter {
-    exists: {
-        field: string;
-    };
-}
-
-export interface NotExistsFilter {
-    notExists: {
-        field: string;
-    };
-}
-
-export interface HasValueFilter {
-    hasValue: {
-        field: string;
-        values: unknown[];
-    };
-}
-
-export interface IdsFilter {
-    ids: {
-        values: string[];
-    };
-}
-
-export interface BooleanFilter {
-    boolean: {
-        must?: Filter | Filter[];
-        mustNot?: Filter | Filter[];
-        should?: Filter | Filter[];
-    };
-}
-
-export type Filter = ExistsFilter | NotExistsFilter | HasValueFilter | IdsFilter | BooleanFilter;
 
 export interface TermSuggestion {
     text: string;
